Added logic to handle invalid twitch usernames

+++ assets/js/options.js
@@ -3,7 +3,7 @@
     "use strict";
 
     chrome.storage.sync.get("user", items => {
-        if (items.user.login) document.querySelector("#user").value = items.user.login;
+        if (items.user) document.querySelector("#user").value = items.user.login;
     });
 
     document.querySelector("#formOptions").addEventListener("submit", event => {
@@ -19,52 +19,48 @@
     document.querySelector(".btn-reset").addEventListener("click", event => {
         chrome.storage.sync.clear(() => {
             chrome.browserAction.disable();
-            chrome.browserAction.setTitle({
-                "title": ""
-            });
-            chrome.browserAction.setBadgeText({
-                "text": ""
-            });
-            showAlert("Data Reset");
+            chrome.browserAction.setTitle({ "title": "" });
+            chrome.browserAction.setBadgeText({ "text": "" });
+            showAlert("Username reset", "success");
         });
     });
 
     const GetUserID = async (login) => {
         let response = await fetch(`https://api.twitch.tv/helix/users?login=${login}`, {
-            headers: {
-                "Client-ID": "haeyonp05j4wiphav3eppivtdsvlyoq"
-            }
+            headers: { "Client-ID": "haeyonp05j4wiphav3eppivtdsvlyoq" }
         });
         let json = await response.json();
 
-        chrome.storage.sync.clear(() => {
-            chrome.storage.sync.set({
-                "user":{
-                    "id": json.data[0].id,
-                    "login":json.data[0].login
-                }
-            }, () => {
-                chrome.runtime.sendMessage({
-                    "action": "resetAlarm"
-                });
+        if (json.data.length > 0) {
+            chrome.storage.sync.clear(() => {
+                chrome.storage.sync.set({
+                    "user": {
+                        "id": json.data[0].id,
+                        "login": json.data[0].login
+                    }
+                }, () => {
+                    chrome.runtime.sendMessage({ "action": "resetAlarm" });
 
-                chrome.browserAction.setTitle({
-                    "title": json.data[0].login
+                    chrome.browserAction.setTitle({ "title": json.data[0].login });
+
+                    showAlert("Username saved", "success");
                 });
-                showAlert("Username Saved");
             });
-        });
+        } else {
+            showAlert(`Username <b>${login}</b> not found`, "danger");
+        }
     };
 
-    const showAlert = message => {
+    const showAlert = (message, alertType) => {
         const alert = document.querySelector(".alert");
+        alert.className = `alert alert-${alertType}`;
 
         alert.innerHTML = message;
         alert.removeAttribute("hidden");
 
         setTimeout(function () {
             alert.setAttribute("hidden", null);
-        }, 2000);
+        }, 4000);
     };
 
 })();
